tools: Separate golang functions & update gh action
