+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"perfect100/utils"
+)
+
+func main() {
+	utils.Achievements()
+}
